requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
Pillow>=10.0.0
qrcode[pil]==7.4.2
psycopg2-binary==2.9.9
whitenoise==6.5.0  # for serving static files